+++ tests/IncludeTest.php
@@ -2,16 +2,16 @@
 
 namespace Spatie\QueryBuilder\Tests;
 
-use ReflectionClass;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
-use Spatie\QueryBuilder\QueryBuilder;
-use Illuminate\Database\Eloquent\Model;
+use ReflectionClass;
 use Spatie\QueryBuilder\AllowedInclude;
 use Spatie\QueryBuilder\Exceptions\InvalidIncludeQuery;
-use Spatie\QueryBuilder\Tests\TestClasses\Models\TestModel;
+use Spatie\QueryBuilder\QueryBuilder;
 use Spatie\QueryBuilder\Tests\TestClasses\Models\MorphModel;
+use Spatie\QueryBuilder\Tests\TestClasses\Models\TestModel;
 
 class IncludeTest extends TestCase
 {
@@ -43,6 +43,20 @@ class IncludeTest extends TestCase
     {
         $models = QueryBuilder::for(TestModel::class, new Request())
             ->allowedIncludes('related-models')
+            ->get();
+
+        $this->assertCount(TestModel::count(), $models);
+    }
+
+    /** @test */
+    public function it_can_handle_empty_includes()
+    {
+        $models = QueryBuilder::for(TestModel::class, new Request())
+            ->allowedIncludes([
+                null,
+                [],
+                '',
+            ])
             ->get();
 
         $this->assertCount(TestModel::count(), $models);
@@ -254,7 +268,8 @@ class IncludeTest extends TestCase
 
         // Based on the following query: TestModel::with('relatedThroughPivotModels')->get();
         // Without where-clause as that differs per Laravel version
-        $this->assertQueryLogContains('select "related_through_pivot_models".*, "pivot_models"."test_model_id" as "pivot_test_model_id", "pivot_models"."related_through_pivot_model_id" as "pivot_related_through_pivot_model_id" from "related_through_pivot_models" inner join "pivot_models" on "related_through_pivot_models"."id" = "pivot_models"."related_through_pivot_model_id"');
+        //dump(DB::getQueryLog());
+        $this->assertQueryLogContains('select `related_through_pivot_models`.*, `pivot_models`.`test_model_id` as `pivot_test_model_id`, `pivot_models`.`related_through_pivot_model_id` as `pivot_related_through_pivot_model_id` from `related_through_pivot_models` inner join `pivot_models` on `related_through_pivot_models`.`id` = `pivot_models`.`related_through_pivot_model_id` where `pivot_models`.`test_model_id` in (1, 2, 3, 4, 5)');
     }
 
     /** @test */
@@ -277,6 +292,26 @@ class IncludeTest extends TestCase
 
         $this->assertEquals(['unknown include'], $exception->unknownIncludes->all());
         $this->assertEquals(['allowed include'], $exception->allowedIncludes->all());
+    }
+
+    /** @test */
+    public function it_can_alias_multiple_allowed_includes()
+    {
+        $request = new Request([
+            'include' => 'relatedModelsCount,relationShipAlias',
+        ]);
+
+        $models = QueryBuilder::for(TestModel::class, $request)
+            ->allowedIncludes([
+                AllowedInclude::count('relatedModelsCount'),
+                AllowedInclude::relationship('relationShipAlias', 'otherRelatedModels'),
+            ])
+            ->get();
+
+        $this->assertRelationLoaded($models, 'otherRelatedModels');
+        $models->each(function ($model) {
+            $this->assertNotNull($model->related_models_count);
+        });
     }
 
     protected function createQueryFromIncludeRequest(string $includes): QueryBuilder