+++ tests/TestClasses/Filters/FiltersTestModels.php
@@ -2,12 +2,12 @@
 
 namespace Spatie\QueryBuilder\Tests\TestClasses\Filters;
 
-use Spatie\QueryBuilder\Filters\Filter;
 use Illuminate\Database\Eloquent\Builder;
+use Spatie\QueryBuilder\Filters\Filter;
 
 class FiltersTestModels implements Filter
 {
-    public function __invoke(Builder $query, $value, string $property) : Builder
+    public function __invoke(Builder $query, $value, string $property): Builder
     {
         return $query->where($property, $value);
     }