+++ tests/TestClasses/Models/ScopeModel.php
@@ -2,8 +2,8 @@
 
 namespace Spatie\QueryBuilder\Tests\TestClasses\Models;
 
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
 
 class ScopeModel extends Model
 {