+++ tests/TestClasses/Models/TestModel.php
@@ -2,13 +2,13 @@
 
 namespace Spatie\QueryBuilder\Tests\TestClasses\Models;
 
-use Illuminate\Support\Carbon;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Illuminate\Support\Carbon;
 
 class TestModel extends Model
 {
@@ -39,12 +39,24 @@ class TestModel extends Model
         return $this->morphMany(MorphModel::class, 'parent');
     }
 
-    public function scopeNamed(Builder $query, string $name) : Builder
+    public function scopeNamed(Builder $query, string $name): Builder
     {
         return $query->where('name', $name);
     }
 
-    public function scopeCreatedBetween(Builder $query, $from, $to) : Builder
+    public function scopeUser(Builder $query, self $user): Builder
+    {
+        return $query->where('id', $user->id);
+    }
+
+    public function scopeUserInfo(Builder $query, self $user, string $name): Builder
+    {
+        return $query
+            ->where('id', $user->id)
+            ->where('name', $name);
+    }
+
+    public function scopeCreatedBetween(Builder $query, $from, $to): Builder
     {
         return $query->whereBetween('created_at', [
             Carbon::parse($from), Carbon::parse($to),